Deprecate PycodestyleBeer.py
Created a new file PyCodeStyleBear.py to deprecate PycodestyleBear.py

Closes: https://github.com/coala/coala-bears/issues/1152
